Adding the right statistoc

+++ src/lib/globals.c
@@ -22,7 +22,7 @@ unsigned int time_diff_micro(struct timeval end, struct timeval start){
   return ((end.tv_sec - start.tv_sec) * 1000000) + (end.tv_usec - start.tv_usec);
 }
 
-unsigned int to_micro(struct timeval tv){
+long long unsigned int to_micro(struct timeval tv){
   return tv.tv_sec* 1000000 + tv.tv_usec;
 }
